Include NAR hash in NAR file name
This ensures that if a store path is garbage-collected and then
recreated with a new NAR hash (i.e. when it's not fully binary
reproducible), clients don't get a hash mismatch error like

  hash mismatch importing path '/nix/store/2v3s58x75nwimfmly8fzkdawn0pvn2ww-foo';
    wanted: sha256:0020y5zzb51ja90in1gx64nn84avmxl4xpbxrba20x6jb2fvy9c5
    got:    sha256:0k98z7m60q29lxmq871dgis7nr03ibkabakd557807fhqx911gav

but instead get

  file 'nar/2v3s58x75nwimfmly8fzkdawn0pvn2ww-0k98z7m60q29lxmq871dgis7nr03ibkabakd557807fhqx911gav.nar' does not exist in binary cache 'http://localhost:5000'

Note that the NAR URLs cannot include *only* the NAR hash, since we
don't have a way to look up store paths by NAR hash.