Support `nix log`
`nix log --store …` expects the remote cache to serve the log output
from a `/log/` endpoint, but currently only Hydra supports that endpoint.
The purpose of this change is to update `nix-serve` to also support the
same endpoint for parity with Hydra, so that `nix log` works when the
cache is hosted using `nix-serve`.